Refacor : Operator 네이밍 전략 수정 및 미사용 객체 주석화 주석화 처리된 객체는 코드 구현후 불필요 판단시 삭제될 예정
