+++ src/main/java/Number.java
@@ -1,6 +1,8 @@
+/*
 public class Number {
 
   public final static String ONLY_NUMBER_REGEX = "^[0-9]+$";
   public final static int FRONT_NUMBER = 0;
   public final static int BEHIND_NUMBER = 1;
 }
+*/
